+++ src/modules/mo-product-dto/product/data-transfer-objects/cluster/admin/cluster-admin-patch-request.dto.ts
@@ -35,7 +35,7 @@ export class ClusterAdminPatchRequestDto extends ClusterAdminCreateRequestDto {
   @Expose()
   provider: ClusterProviderEnum;
 
-  @Transform(({ value }) => value ?? ClusterBuildServerTypeEnum.MO_AZURE)
+  @Transform(({ value }) => value ?? ClusterBuildServerTypeEnum.IN_CLUSTER)
   @Expose()
   buildServerType: ClusterBuildServerTypeEnum;