+++ src/modules/mo-product-dto/product/data-transfer-objects/cluster/cluster.dto.ts
@@ -30,7 +30,7 @@ export class ClusterDto extends BaseEntityDto {
   @Expose()
   provider: ClusterProviderEnum;
 
-  @Transform(({ value }) => value ?? ClusterBuildServerTypeEnum.MO_AZURE)
+  @Transform(({ value }) => value ?? ClusterBuildServerTypeEnum.IN_CLUSTER)
   @Expose()
   buildServerType: ClusterBuildServerTypeEnum;