+++ src/modules/mo-project-dto/data-transfer-objects/project-namespace-service-container/project-namespace-service-container.dto.ts
@@ -14,6 +14,7 @@ import { OriginTrafficDto } from '../traffic';
 import { KubernetesPublicEventDto } from '../../../mo-kubernetes';
 import { ContainerTypeEnum } from '../../enums';
 import { ProjectNamespaceServiceContainerKubernetesSettingsDto } from './project-namespace-service-container-kubernetes-settings.dto';
+import { AppContainerDto } from '../../../mo-app-library-dto/data-transfer-objects/app-container/app-container.dto';
 
 export class ProjectNamespaceServiceContainerDto extends BaseEntityDto {
   @Transform(({ value }) =>
@@ -94,13 +95,13 @@ export class ProjectNamespaceServiceContainerDto extends BaseEntityDto {
   @Expose()
   ephemeralStorage: EphemeralStorageDto;
 
-  @Type(() => OriginTrafficDto)
-  @Expose()
-  traffic: OriginTrafficDto;
-
   // deployment message
   @Transform(({ value }) => (value && isArray(value) ? value : []))
   @Type(() => KubernetesPublicEventDto)
   @Expose()
   messages: KubernetesPublicEventDto[];
+
+  @Type(() => AppContainerDto)
+  @Expose()
+  appContainer: AppContainerDto;
 }